refactor(observability): split health.py into contracts, renderers, and checker
health.py was the largest first-party module (521 lines) bundling three
concerns behind one interface. Now:
- health_contracts.py — status vocabulary, probe/component records, payload
  TypedDicts, overall_health_status, iso_z (no I/O)
- health_renderers.py — detailed JSON payload + Prometheus text rendering
- health.py — HealthChecker: live dependency probes and resource sampling

Consumers import from the owning module; no re-export shim.
