+++ backend/tests/unit/test_package_contracts.py
@@ -58,7 +58,7 @@ def test_api_infrastructure_has_explicit_package_ownership() -> None:
     api_root = backend_root / "stream_sniper/api"
     expected_modules = {
         "caching": {"__init__.py", "cache.py", "cache_warmup.py", "model_cache.py", "rollup_version.py"},
-        "observability": {"__init__.py", "health.py", "monitoring.py"},
+        "observability": {"__init__.py", "health.py", "health_contracts.py", "health_renderers.py", "monitoring.py"},
         "security": {"__init__.py", "auth.py", "auth_models.py", "rate_limiter.py"},
         "transport": {"__init__.py", "export_utils.py", "models.py"},
     }
